+++ Builder/Managers/Widgets.php
@@ -214,61 +214,6 @@ class Widgets
     }
 
     /**
-     * Retrieve inline editing configuration.
-     *
-     * Returns general inline editing configurations like toolbar types etc.
-     *
-     *
-     * @return array {
-     *     Inline editing configuration.
-     *
-     *     @type array $toolbar {
-     *         Toolbar types and the actions each toolbar includes.
-     *         Note: Wysiwyg controls uses the advanced toolbar, textarea controls
-     *         uses the basic toolbar and text controls has no toolbar.
-     *
-     *         @type array $basic    Basic actions included in the edit tool.
-     *         @type array $advanced Advanced actions included in the edit tool.
-     *     }
-     * }
-     */
-    public function getInlineEditingConfig()
-    {
-        $basic_tools = [
-            'bold',
-            'underline',
-            'italic',
-        ];
-
-        $advanced_tools = array_merge($basic_tools, [
-            'createlink',
-            'unlink',
-            'h1' => [
-                'h1',
-                'h2',
-                'h3',
-                'h4',
-                'h5',
-                'h6',
-                'p',
-                'blockquote',
-                'pre',
-            ],
-            'list' => [
-                'insertOrderedList',
-                'insertUnorderedList',
-            ],
-        ]);
-
-        return [
-            'toolbar' => [
-                'basic' => $basic_tools,
-                'advanced' => $advanced_tools,
-            ],
-        ];
-    }
-
-    /**
      * Widgets manager constructor.
      *
      * Initializing SagoTheme widgets manager.